代码提交

+++ server-config-manager/server-config-manger-dao/src/main/java/com/server/config/manger/dao/hibernate/ResourceDao.java
@@ -12,7 +12,7 @@ import java.util.List;
  */
 @Repository
 @Transactional
-public class ResourceDao extends BaseDao<Resources,Long> {
+public class ResourceDao extends BaseDao<Resources,Integer> {
 
     /**
      * 根据code获取资源
